+++ lib/methods.ts
@@ -126,8 +126,6 @@ export {
   defineStep as Given,
   defineStep as When,
   defineStep as Then,
-  defineStep as And,
-  defineStep as But,
   defineStep,
   runStepDefininition as Step,
   defineParameterType,